Bug Fixes and Refactoring
Fixed a bug where the app would crash when the user logs out. I'm still not sure what was causing this exactly, but dismissing the UserSettingsView (which is now a full screen cover) before presenting the LoginView seems to have fixed it. Also, the background color used throughout the app has been refactored into its own BackgroundColor view. Finally, view state management was added to SearchView to make it clearer to the user when they perform a search that has no results