More Unit Tests and Bug Fixes
Added more Unit Tests for testing computed properties in model data. Also fixed a bug where an invalid view state message would be shown after sending an invite to someone to join a band. Finally, more example data was added to FirebaseEmulatorData, and the BandMember and BandInvite no longer use the @DocumentID and @ServerTimestamp to make testing easier. Instead, they have new properties that are written to when objects of those types are created. This will be applied to the remainder of types that use these property wrappers as well.