Added Extensions Emulator and More Tests
Tests were added to EditImageViewModel and HomeViewModel. Also, Firebase Extensions Emulator was set up and appears to be working correctly. Because this is the case, I've commented out the lines in TheSamePageApp.swift that were causing the emulator to only run during tests. Now, it always runs to make testing easier both in written tests and in tests that can be performed by using the app.