Added More Tests and Miscellaneous Changes
Made MemberSearchView in BandProfileView a fullScreenCover instead of a NavigationLink. Also fixed strange sheet dismiss animation behavior in LinkTab in BandProfileView, and BandProfileViewModel now includes individual methods for fetching a band's members, shows, and links. These methods are called in a callOnAppearMethods() method. This was done because this functionality was previously included in addBandListener(), which didn't make sense because changes to a band's members, shows, and links will not trigger changes for updating the band listener. Finally, tests for LoginViewModel, SignUpViewModel, OtherUserProfileViewModel, SendBandInviteViewModel, BandProfileViewModel, and RootViewModel were added.