Added More Tests and Fixed User Image Editing Bug
Finished tests for ShowDetailsViewModel and AddBacklineViewModel. Also fixed a bug where updating a user's image twice back to back would throw an error. This was happening because the LoggedInUserController's loggedInUser and profileImageUrl properties weren't getting updated with the new image. Now, the EditImageView updates this info whenever it's dismissed. Finally, EditImageView is now presented as a full screen cover for updating a user's profile image. In a later commit, EditImageView will always be presented as a fullScreenCover.