Modified EditImageView and SettingsView Functionality
EditImageView and all SettingsViews are now presented in a .fullScreenCover. These changes were made so that it would be more difficult for the user to interrupt important operations while they were happening. For example, it is now impossible to move to another tab while an image update is occuring, which will make it harder for the user to do other things that may interfere with the image update.