Added New NoDataFoundMessageWithButton, Modified Modal Presentations, Design Changes, and Fixed Bugs
Fixed a bug where signing into one user's account, and then signing into a different account would cause the old user's profile image to appear in EditImageView. Also, moved buttons for sending band invites and show invites to the OtherUserProfileView and BandProfileView toolbars, respectively. The chat button in ShowDetailsView was also moved to the toolbar. This made it easier to match the header design of BandProfileView and ShowDetailsView, which has been modified with Label views that show the most important show info at a glance. Additionally, NoDataFoundMessageWithButton is a new view that's been added in various spots to make it easier to tell the user that no data has been found for a view they're looking at. This view may include a button to allow the user to add whatever data couldn't be found. Finally, the presentation of various views has been made more consistent throughout the app. For instance, any time an AddEdit view is presented modally, it's presented with a .fullScreenCover. Additionally, these views are presented as NavigationLinks inside SettingsViews because the SettingsViews are now presented modally.