Added Functionality for Adding Your Own Band to a Show You're Hosting
Added AddMyBandToShowView. This view allows the user to add any bands of which they are an admin to any shows they are hosting. Tests for this AddMyBandToShowViewModel were also added. Finally, SendBandInviteView now fetches the user's bands in its .task method. Before, it was doing this in the SendBandInviteViewModel's init, which was not ideal because it was less predictable.