Delete Image Functionality Added and View Refactoring
Deleting an image is now possible in BandProfileView, ShowDetailsView, and LoggedInUserProfileView. Tests have been written for this functionality in the first two views listed, but not yet for LoggedInUserProfileView because LoggedInUserController tests haven't yet been written. All three of these views have refactored versions of their profile images to make it easier to enable their button-like functionality when the logged in user meets certain conditions.