Deleting Backline Items is Now Possible
It is now possible to delete a backline item from a show. However, this functionality has a bug that keeps the UI from updating when a backline item is deleted. This will need to be resolved with a new Backline protocol along with a new AnyBacklineItem wrapper. This functionality will be implemented in a later commit.