Modified Navigation Flow in ProfileViews and ShowDetailsView
Bands selected in ShowDetailsView and UserProfileViews, Shows selected in BandProfileView, and users selected in BandProfileView now appear in a full screen cover, as opposed to a NavigationLink. This was done to make it look neater when a trash can/delete icon gets added to these views to allow for removing bands from a show and members from a band. This icon appearing next to the NavigationLink chevron appeared too cluttered, but in a future update when the minimum target version is incremented, this will be revisted since lists with swipe to delete will be implemented. Also fixed an issue where the trash can/delete icon was not appearing centered relative to other text in a row.