Altered Navigation Flow and Removing a Band From a Show Is Now Possible
Navigation flow now makes more sense in that NavigationLink is used when presenting a view that is closely related to the one that came before it. For instance, chats and settings views are no longer full screen covers. Also, it is not possible for a show host to remove a band from a show's lineup. This removes that band from the show's bandIds property and the show's participants collection. It also removes the band's members' UIDs from the show's chat's participantUids property. In a future commit, this will also remove anything that was backlined by any of the members in the removed band.