A Band Can Now Leave a Show
It is now possible for the admin of a band to remove their band from a show via ShowDetailsView's Lineup tab. In order to make this work, a property was added to the ShowParticipant struct, which is bandAdminUid. This property, combined with the new ShowParticipant bandAdminIsLoggedInUser computed property, makes it easier to tell whether or not the Leave Show button should be displayed in ShowDetailsView's Lineup tab. Tests were also written for all of the above functionality.