Removing a BandMember from a Band Is Now Possible and Bug Fixes
It is now possible for a band admin to remove a band member from the band via BandProfileView's Members tab. Also fixed a bug that would sometime cause the incorrect profile to be shown when one was selected via ShowDetailsView's Lineup tab. To fix this bug, the .fullScreenCover is now triggered via an optional value instead of a boolean. This pattern was adapted in BandProfileView and the UserProfileViews to avoid a similar bug in those views. Also fixed a bug where the dismissal of the BandProfileView in ShowDetailsView's Lineup tab was not animated. Finally, modified the workflow for various row views throughout the app. These row views are usually passed an index so that they can receive realtime updates from their associated view models. However, sometimes referencing the index that they are passed would cause a crash due to the unexpected times at which SwiftUI would evaluate the index of a given array. This was fixed by ensuring that the array in question contains the index that was passed to the row before actually referencing that index in the array and risking causing a crash. If it does, the row is created and displayed. If it doesn't, the row is never created. Also, the functionality for leaving a band was taken away from BandSettingsView. That's because the user will now be able to leave the band via their profile view instead to be more consistent with the rest of the app.