Deleting a Band is Now Possible and New ViewState Case
Added .dataDeleted ViewState case for when a view is attempting to display data that was deleted by the user. This has only been implemented in BandProfileView so far, but it will be implemented in other views, such as ShowDetailsView, in a future commit. Also, deleting a band is now possible from BandSettingsView. This functionality has new tests for it as well.