Implemented Reset Password Functionality
It is now possible for the user to reset their password if they forgot it by clicking the Forgot Your Password? button in LogInView. This will present them with ForgotPasswordView, where they will enter an email address to which Firebase will send a password reset email. When they receive this email, they will click a link that will redirect them to a site they can use to change their password. When their password has been reset, they can click a continue button to get directed back to the app. In order for this functionality to work, the Xcode project was updated with the Associated Domains capability. This makes it possible for the app to be opened via the thesamepage.page.link Dynamic Link set up in Firebase. A new custom URL Type was also added to the project for the same reason.