New Username Creation Flow and Bug Fixes
A new step was added to onboarding that forces the user to create a username after they've created their account, but before they can use their account. This had to be done because every username must be unique, but it is not possible to check for uniqueness unless the user is signed in due to security rules. Bugs were also fixed that would allow the user to use the app if they had an unverified email address or if they had not yet set up an account. Another bug was fixed where the user's location information wasn't set before nearby shows were fetched in HomeView. This would result in a .dataLoading view state that would never resolve. To fix this, a notification is now posted when the user's location is set in LocationController. This same notification is observed in HomeView, and when the observer receives the notification the shows are fetched, ensuring that the search for nearby shows doesn't begin until the user's location is known.