Implemented UserSettingsNavigationViewModel in UserSettingsView Navigation Flow
NavigationStack, .navigationDestination, and NavigationPath have all been implemented in the navigation flow in UserSettingsView. This was done because the NavigationLink isActive property was deprecated in iOS 16 and the target of the project has been bumped up. This change will be made to other Views as well. Also added MockBackButton and MockNavigationLinkRow views to make it easier to recreate the look of back button and NavigationLink List rows, respectively, in the app when more custom implementations are needed.