Changed Name of HomeView and Converted TextEditors to Vertical TextFields
HomeView is now FindShowsView and the tab item for HomeView now says Find Shows with a different SF Symbol. This was done to more efficiently convey the use of the view without navigating to it. Finally, all TextEditors were converted into TextFields with a .vertical axis property. This was done because the minimum target was bumped to iOS 16, which is the only iOS version with an axis property for TextField, and because TextEditor tends to appear glitchy (shifts slightly when focused and padding around text is never even).