Choosing a New Band Admin is Now Possible
It is now possible for a band admin to choose a new admin for their band. Currently, only other band members can be chosen as the new band admin. However, this may change in the future so that anybody can become a band admin regardless of whether or not they're already in the band.