List Styling Changes and Band Links Can Now Be Deleted
Changed the styling for destructive actions in ShowDetailsView and ProfileViews. Before, trash SF Symbols were always shown to allow for a band to be deleted from a show, a user to be removed from a band, etc. Now, though, an ellipses SF symbol is shown that presents a Menu view to give the user the option to take action on an object. This is better stylistically and it will also make it easier to make it possible to edit these objects in a future commit. The only issue with this functionality is that it can sometimes delete the incorrect object from the array it's referencing. I'm not quite sure why this is happening, but in a future commit I will attempt to fix this by referencing the object that is to be deleted in the MenuButton views the same way as I'm referencing it in RowViews via its index.