Added Backline Protocol, added AnyBackline Wrapper, and Modified Backline System
BacklineItem and DrumKitBacklineItem now conform to the Backline protocol. Additionally, the Backline tab in ShowDetailsView now iterates through an array of AnyBackline, a new property wrapper that makes it easier for a ForEach or List to iterate through any objects that conform to Backline. Also, an Ellipses icon was added to each ShowBacklineRow to allow for a Backline item to be deleted. This deletion now updates the UI in realtime. Because all of the items in the Backline tab are iterating through the same array, the sections for each BacklineItem and DrumKitBacklineItem type have been removed from the UI. Now all Backline is shown together in a sorted list, which required certain types of gear to be renamed so that the list would be clearer (Electric Guitar Cab instead of just Guitar Cab, for example). Necessary changes to facilitate this funcitionality have also been added to AddBacklineView.