Fixed Unexpected Delete Behavior Bug
In previous commits, items were not properly being deleted in the ProfileViews and ShowDetailsView. For instance, when deleting a ShowParticipant from a show, it was possible that the incorrect ShowParticipant would get deleted from the show. This was happening because all of the delete confirmation alerts for each row of each deletable object were being triggered by the same boolean in their respective view model. As a result, it was never possible to predict which row's alert was being presented because they were all trying to present themselves at the same time. This was fixed by moving the booleans that trigger the delete confirmation alerts into the MenuButton for each deletable object's row. This ensures that each alert for each row has a unique boolean that will only get triggered by that particular row.