Fixed DrumKitBacklineItem Included Kit Pieces List Bug
In ShowDetailsView's Backline tab, DrumKitBackLineItems are displayed with a Full Kit title and a description of all the included kit pieces, alongside any other notes written by the user, as a subtitle. This functionality was quite messy in a computed property called details on DrumKitBacklineItem. It was also buggy because the list's punctuation would look strange if all kit pieces weren't being backlined. This property was replaced with includedKitPiecesFormattedList, which uses a ListFormatter to create a list of all the included kit pieces listed in the new includedKitPieces array. The includedKitPieces property on DrumKitBacklineItem is also new, and it's replaced the kickIsIncluded, snareIsIncluded, etc. properties. This array is created by a new method in AddBacklineViewModel that evaluates all the booleans that are set in the AddBacklineView's form, and then adds strings to represent each kit piece accordingly.