Implemented alreadyHappened Property On Show and Bug Fixs
The Show object now has an alreadyHappened property that evaluates whether or not a show's date is in the past. This has been implemented throughout the app and it is now replacing every place that used a comparison operator in relation to Date.now, which was a more confusing way of achieving the same functionality. Also, a chevronColor property was added to MockNavigationLinkRow to allow for the chevron to be red when used in conjunction with a .destructive role. Finally, fixed a bug where LeadingBandsAndShowsList showed all shows, and not just upcoming ones.