Email Address Changing is Now Possible
It is now possible for the user to change their email address on The Same Page. When they do so, they automatically get logged out and they must verify their new email address before signing in again. Also, UsernameTextField was renamed to CustomTextField and a keyboardType property was added to it. This was done because the same functionality for UsernameTextField was very useful for email address TextFields as well, so this avoids the need to complete a separate view for the two different types of TextFields.