Tightened Up Button and Dismissal Disabling in Sheets During Async Operations
Added .interactiveDismissDisabled modifiers to all views presented as sheets. This modifier keeps the user from being able to swipe the view down while an important async task is happening. .navigationBarBackButtonIsHidden and .disabled modifiers were also applied in various views that perform async operations to make it more difficult for the user to interfere with the operation.