Bands and Shows Are Now Viewable from NotificationsView
A Band Info button now appears on BandInvite and ShowApplication notifications in NotificationsView. Similarly, a Show Info button appears on ShowInvite notifications in NotificationsView. This can be tapped to view more information about the band or show that the notification pertains to. This was done to make it easier to get more context on a notification. In order to make this work, ShowDetailsViewModel and BandProfileViewModel can now be initialized with a document ID property.