Push Notifications Get Sent for New BandInvites, ShowInvites, and ShowApplications
A new Firebase Cloud Function notifyNewNotification was added that allows for a push notification to get sent to a user when they receive a new ShowInvite, ShowApplication or BandInvite. This Cloud Function listens to all notifications collections for all users and sends the notification when a new document is created in a notifications collection.