Show Details Can Now Be Edited for Show In The Past
It is now possible to edit details for a show that already took place. This was done because creating a show that was taking place the same day would cause a bug where the should could not be edited. However, it is not possible to cancel or assign a new host to a show that already took place.