Removed All Deprecated NavigationLink Calls
All NavigationLink usages that utilize the now-deprecated isActive property have been replaced with the appropriate NavigationStack functionality. The only place this change needed to happen was during the sign up process, so an OnboardingNavigationViewModel was created to facilitate this. This also solves a bug where the user wouldn't get pushed back to LoginView after creating a username during onboarding. This was happening because the isActive NavigationLink property was being used in conjunction with a NavigationStack.