Tapping a New Message Notification Opens Chat Within App
AppOpenedViaNotificationController was added to make it easier to allow for views to be presented when the app is opened via a notification. For now, this only allows the app to present ConversationView when a new message notification is tapped, but this functionality will be extended to BandInvite, ShowInvite, and ShowApplication notifications too. To allow for this functionality, ConversationView can now be initialized with a chatId property. This functionality is a bit messier than I would like for it to be, but this will be cleaned up once chats are not exclusive to shows.