.error ViewStates No Longer Present an EmptyView
To create more continuity in the app and avoid strange visual glitches, .error ViewStates are no longer presented with an EmptyView. Instead, they work alongside .dataLoaded or .workCompleted, which will make the error message look better when it's presented and it will also make it easier to allow the user to try again via .refreshable.