FindShowsView, MyHostedShowsView, and MyPlayingShowsView Only Show Upcoming Shows and Bug Fixes
In addition to only showing upcoming shows in FindShowsView, MyHostedShowsView, and MyPlayingShowsView, the verbage was also changed in these views to indicate that the user will only see upcoming shows in these views. Additionally, it is now impossible for a user to accept a ShowInvite or ShowApplication for a band that is already playing the show that the invite/application pertains to. Similarly, it is now impossible for the user to accept a BandInvite for a band that they already play in.