Changed .alreadyHappened Property On Show
A Show's .alreadyHappened property now only considers a show to have already happened if it is at least midnight on the day after the show. This was done to fix a bug where a show was not shown as upcoming and thus could not be edited if the current date was the same day as the show. That bug was happening because the time that the show was created was assigned to the show's time in its date property, making it so that if the show was created on the same day that it was set to happen then its .alreadyHappened property would instantly be true.