Fixed AddMyBandToShowViewModel Bug and Started Implementing FCM Token Maintenance
AddMyBandToShowViewModel now fetches all bands of which the user is an admin, instead of fetching bands of which they're a member and then filtering it for bands of which they are an admin. Also, methods were implemented in DatabaseService to delete and update a user's FCM token. This will be helpful for the next commit, where there will be changes to make FCM token maintenance a lot easier.