Optimized FCM Token Handling for ChatMessages and FCM Token Updating
The participantFcmTokens property has been removed from Chat. This was done because maintaining updated FCM tokens in this property would've been very difficult. Instead of this, ChatMessages populate their recipientFcmTokens property by fetching the FCM token associated with the users in their Chat's participantUids property. This ensures that the ChatMessages always have the most up to date FCM Tokens. Also, the logged in user's FCM Token is now deleted when they log out, restored when they logged in, and updated when the open the app when they were already logged in. This was done to make sure that the logged in user always has the most up to date FCM Token associated with their entry in Firestore.