Optimized FCM Token Behavior for Accepting Invites and Applications and Bug Fix
The senderFcmToken property was removed from the UserNotification protocol. This was done because the user that sent the invite or application could have an updated FCM Token by the time the application or invite is accepted. To work around this issue, a senderUid property was added to the UserNotification protocol. By associating the sender's UID with all UserNotifications, it is now possible for the sender's most up to date FCM token to get fetched from their entry in the users collection of Firestore. Also, a bug was fixed where removing a band from a show's lineup would not remove the band's admin from the show's participantUids array if the band's admin was not also a member of the band.