Updated LoggedInUserProfileView Design
LoggedInUserProfileView now has a segmented Picker that allows the user to see their shows and bands. A section was also added to LoggedInUserProfileHeader that shows the number of bands and shows the user is involved in. To facilitate this, methods were added to LoggedInUserController and DatabaseServiec that fetch all of the shows and bands that the user is involved in, meaning any show or band that the user runs or is involved in in any other way. This design will also be added to OtherUserProfileView in a future commit