Started Implementing Dynamic Links and Notification Bug Fixes
The beginning stages of Dynamic Link implementation are in the TheSamePageApp struct. Also, TheSamePageApp was modified to fix a bug where tapping a notification wouldn't always present the view related to the notification. The fix for this bug is quite ugly and inelegant, but it doesn't appear there's a better way to implement this functionality.