Fixed Bugs that Occurred When Tapping Notifications
Previously, tapping a notification while a sheet was already being presented would not consistently result in showing the content that pertained to the notification to the user, especially when the notification was tapped while another view showing info pertaining another notification was already being presented. This has been fixed. Additionally, a bug was fixed where the user would receive a notification about a new message in a chat while they were already looking at the chat. This was fixed by adding a currentViewerUids property to Chat, which is updated when ConversationView appears, when ConversationView disappears, and when the scenePhase changes. Extensions were added to assist with scenePhase updates because switching on @Environment(\.scenePhase) was not consistently working. When a message is sent, a notification is not sent to any user whose UID is included in the currentViewerUids property.