Fixed Bug When Opening App From Notification
Previously, opening an accepted application/invite notification immediately after opening a new message notification would not work. This has been fixed. However, a bug still remains where opening a new message notification immediately after opening a new message notification for a different chat does not show the new chat. This bug shouldn't be an issue for users most of the time and I couldn't figure out a solution, so it will be deferred for now.