Completed Setup for Opening Band and User Dynamic Links and Added Shareable Protocol
The app is now capable of accepting Band and User Dynamic Links, but cannot act on them yet. OtherUserProfileView still needs to be able to accept a UID as an argument before this will be possible, and BandProfileView and OtherUserProfileView still need a share button. Also, the Shareable protocol was added and the DynamicLinkType enum was added. These types make it easier to use one method for all Dynamic Link creation.