Bands Are Now Shareable
Dynamic Link functionality has been integrated for sharing band Dynamic Links and opening band Dynamic Links. Also, all .navigationBarTrailing toolbar items in BandProfileView and ShowDetailsView are now presented within an EllipsesMenuIcon. This was done because it allows the share button to load and appear in the background within the Menu without a visual delay. Before, there would be a visible delay while the shortened Dynamic Link was generated in the background, and it would look strange when the Share button appeared after the rest of the view. A new property was also added to ShowDetailsViewModel and BandProfileViewModel that dictates whether or not this EllipsesMenuIcon Menu should be shown in the first place.