Other Users are Shareable and Link Sharing Bug Fix
It is now possible to share other users that are on The Same Page. Also, a bug was fixed that was causing Dynamic Links to become altered and malformed when shared via iMessage. It appears the only way to work around this is to prepend a period onto the link so that iMessage treats it as tappable plain text instead of as a URL. This was done for the band, show, and other user sharing functionality, and it appears that the link still works just fine on other platforms as well. Also, a bug was fixed where, in some places, the old username property was still being referenced on User, even though this property's name is now name instead of username.