Added Styling to LoggedInUserProfileView Chat Button
Added Badge, which is a new view that displays at the top right corner of the chat button in LoggedInUserProfileView. Badge displays the number of unread chats by looking at all of the user's Chats and filtering out the ones where the user is not included in the chat's upToDateParticipantUids array.