Restructured ConversationViewModel Initialization
It is now possible to pass a Chat object to ConversationViewModel for situations where all the info about a given chat is already known. Also, the overall structure for initializing a chat in ConversationViewModel in certain situations has been refactored and is more readable. It should also be much easier to expand upon when chats between individual users are possible.