Keyboards Can be Dismissed on Forms and Updated Notification Text
New chat message notifications for shows now include the name of the show that the chat message refers to, along with the name of the person that sent the message. Also, all keyboards for all forms can now be dismissed by swiping down. This was done with the .scrollDismissesKeyboard(.interactively) modifier.