Fixed Logging Out and Deleting Account Bugs
Fixed a bug where sometimes logging out wouldn't work as expected. Also, moving between the emulator and real database now works as expected and doesn't require resetting the simulator, and opening the app after you're account's been deleted now presents the LoginView as expected. This was done by adding currentUserIsInvalid = true to all guard else statements in methods that are called on app launch. Finally, a ProgressView is shown in RootView when LoggedInUserController.loggedInUser is nil.