Search Improvements and Removed memberFcmTokens Property from Band
A ProgressView is now shown when search operations occur. Additionally, the memberFcmTokens property was removed from Band. This was there because it was aiding in adding a user's FCM Token to a chat's participantFcmTokens property, but since that property was removed from Chat in a previous commit, it's no longer necessary on Band.