Cannot Tap Logged In User's Username in ChatInfoView
It is no longer possible for one to tap their own username in ChatInfoView. This was done to make the view's behavior more consistent with the rest of the app, where it also isn't possible for one to tap a row for their own username. Additionally, ChatInfoView now shows the host of a show in a slightly different way that will make it easier to integrate group chats at a later time.