Improved FindShowView Behavior for When User Location Is Not Known
When the user has denied location permissions, they are now shown a message on FindShowsView telling them that they cannot search for nearby shows. This message is accompanied by a button that, when tapped, takes the user directly to the page of the Settings app that will allow them to change the location permission status.