Started Security Rules and Added Show Property
Added chatId property to Show to make querying a show chat easier. If I didn't do this, I would have to query all of the user's chats for a second time in the app's lifecycle every time a show chat got opened, which is obviously not ideal. Also started incorporating Firestore rules and made miscellaneous changes to make this work.