AddEditBandView and ShowDetailsView Bug Fixes
AddEditBandView now re-enables its buttons when a .error ViewState is set. Also, ShowDetailsView fixed a bug where adding backline to a show that already happened was possible. It was also possible to add backline to a show you weren't participating in if the show already had backline. Also, it is no longer possible to edit the times in the ShowTimeTab if the show already happened.