Bug Fixes
The Firestore FieldValue.arrayUnion/arrayRemove properties are no longer being used when adding or removing a user from a show or a chat. This is because these methods do not allow the array that they're updating to have any duplicate values, so if a user was in two bands on a show or chat and one of those bands left, they would be removed from the show and chat even though they are still part of one band that was a part of the show and chat. Additionally, deleting a band, cancelling a show, and deleting a user account now deletes any images that are associated with that show, user, or band. The functionality behind deleting a user account was also modified to work with Security Rules. Before, AuthController.logOut() was being called after a user's account was deleted. This was not necessary and it was causing issues with Security Rules, so this was removed. Finally, a bug was fixed where a user could not delete their account if they were the host of any shows at any time. Now, the user can delete their account as long as they're not hosting any shows that have not happened yet and as long as they are not the admin of any bands.