Fixed a Chat Bug
Previously, in order to determine the FCM Token of each user that should receive a notification about a chat message, each user in the chat's participantUids array would be iterated over and then their FCM Token would be fetched and used to send the notification. However, this did not account for instances where a chat has the same uid listed multiple times in its participantUids property. Now, the code loops over a Set of all the chat's participantUids instead of an array to find the FCM Token for each user. This ensures that each user's FCM token will only be fetched once, and it will avoid the undesired behavior of sending a user two notifications about the same message if their UID is in the chat twice.