Fixed NotificationsView Formatting Issues
All of the icons in NotificationRow are now formatted to the same size so that different icons don't cause the text in the row to be indented by different amounts. Additionally, the layout of the buttons in NotificationRow was changed so that they take up less horizontal space. The accept and decline buttons are above the info button.