Fixed LoggedInUserProfileView Bug
Fixed a bug where a band would only be shown in the Bands tab of LoggedInUserProfileView if the logged in user was playing in the band. This was happening because the list was iterating over the wrong array. Now, the logged in user can see all their bands here, even the ones that they are the admin of but are not playing in.