Added More Social Icons and Dynamic Link Improvements
There is now an icon for each of the social media options that appear when adding a link to a band's profile. A band now also has the option to add their website to their links. Additionally, Dynamic Links created programmatically now do not include an app store ID. This was done because the app will not have an app store ID until after it's released, so at the initial release all dynamic links will take you to the JMT website if you do not have the app installed. The specific page that you will be taken to on the JMT website has not been created yet. Finally, the preview image for all dynamic links now points to the correct app icon, which is now being hosted in Firebase Storage.