Changed AddEditShowView Terminology and New JMT Website Endpoint for Dynamic Links
A page has been created at julianmichaeltechnologies.com/thesamepage that hosts all the info about the app. DynamicLinkController has been modified to use this new enpoint, /thesamepage, when creating all dynamic links so that any user that opens an app link on an Android phone, laptop, computer, or tablet will get taken to a page that makes sense. After release, all programmatically generated dynamic links will take the user to julianmichaeltechnologies.com/accountmodificationlandingpage if a user clicks it from an Android phone, laptop, computer, or tablet. This makes sense because, at release, I won't have an appStoreId property for the dynamic link, so users would get taken to that page even if they're on iOS. Also, AddEditShowView now says 'Admission is free' instead of 'Show is free' in the Form for the sake of clarity.